Update to new preopt pass name
