+++ filetests/simple_preopt/div_by_const_power_of_2.clif
@@ -1,4 +1,4 @@
-test preopt
+test simple_preopt
 target i686 baseline
 
 ; -------- U32 --------
