+++ filetests/simple_preopt/simplify.clif
@@ -1,4 +1,4 @@
-test preopt
+test simple_preopt
 target i686
 
 function %iadd_imm(i32) -> i32 {
