+++ lib/filetests/src/test_simple_preopt.rs
@@ -1,50 +1,44 @@
-//! Test command for testing the constant folding pass.
-//!
-//! The `dce` test command runs each function through the constant folding pass after ensuring
-//! that all instructions are legal for the target.
+//! Test command for testing the preopt pass.
 //!
 //! The resulting function is sent to `filecheck`.
 
 use cranelift_codegen;
 use cranelift_codegen::ir::Function;
 use cranelift_codegen::print_errors::pretty_error;
-use cranelift_preopt::optimize;
 use cranelift_reader::TestCommand;
 use std::borrow::Cow;
 use subtest::{run_filecheck, Context, SubTest, SubtestResult};
 
-struct TestOptimize;
+struct TestSimplePreopt;
 
 pub fn subtest(parsed: &TestCommand) -> SubtestResult<Box<SubTest>> {
-    assert_eq!(parsed.command, "optimize");
+    assert_eq!(parsed.command, "simple_preopt");
     if !parsed.options.is_empty() {
         Err(format!("No options allowed on {}", parsed))
     } else {
-        Ok(Box::new(TestOptimize))
+        Ok(Box::new(TestSimplePreopt))
     }
 }
 
-impl SubTest for TestOptimize {
+impl SubTest for TestSimplePreopt {
     fn name(&self) -> &'static str {
-        "optimize"
+        "simple_preopt"
     }
 
     fn is_mutating(&self) -> bool {
         true
     }
 
-    fn needs_isa(&self) -> bool {
-        true
-    }
-
     fn run(&self, func: Cow<Function>, context: &Context) -> SubtestResult<()> {
-        let isa = context.isa.expect("compile needs an ISA");
         let mut comp_ctx = cranelift_codegen::Context::for_function(func.into_owned());
+        let isa = context.isa.expect("preopt needs an ISA");
 
-        optimize(&mut comp_ctx, isa)
+        comp_ctx.flowgraph();
+        comp_ctx
+            .preopt(isa)
             .map_err(|e| pretty_error(&comp_ctx.func, context.isa, Into::into(e)))?;
 
-        let text = comp_ctx.func.display(context.isa).to_string();
+        let text = &comp_ctx.func.display(isa).to_string();
         run_filecheck(&text, context)
     }
 }
